Split Android AI bootstrap coordinator
